Move hooks and services to dedicated files
