+++ src/services/helpers/scopePermissions/hooks/resolveScope.js
@@ -1,0 +1,14 @@
+const resolveScope = (context) => {
+	try {
+		const scopeName = context.path.match(/^\/?(\w+)\//)[1];
+		const { scopeId } = context.params.route;
+
+		return { scopeName, scopeId };
+	} catch (e) {
+		return {};
+	}
+};
+
+module.exports = {
+	resolveScope,
+};
